HAT moving all static OpWrapper methods to OpTk
