+++ hat/backends/jextracted/shared/src/main/java/hat/backend/jextracted/C99JExtractedBackend.java
@@ -33,6 +33,7 @@ import hat.buffer.KernelContext;
 import hat.callgraph.KernelCallGraph;
 import hat.ifacemapper.BoundSchema;
 import hat.ifacemapper.Schema;
+import hat.optools.OpTk;
 
 import java.util.Arrays;
 import java.util.HashMap;
@@ -98,7 +99,7 @@ public abstract class C99JExtractedBackend extends JExtractedBackend {
         System.out.println("Original");
         System.out.println(kernelCallGraph.entrypoint.funcOpWrapper().op.toText());
         System.out.println("Lowered");
-        System.out.println(kernelCallGraph.entrypoint.funcOpWrapper().lower().op.toText());
+        System.out.println(OpTk.lower(kernelCallGraph.computeContext.accelerator.lookup,kernelCallGraph.entrypoint.funcOpWrapper().op).op.toText());
 
         return builder.toString();
     }
