+++ hat/core/src/main/java/hat/codebuilders/C99HATKernelBuilder.java
@@ -32,7 +32,7 @@ import hat.callgraph.KernelCallGraph;
 import hat.callgraph.KernelEntrypoint;
 import hat.ifacemapper.MappableIface;
 import hat.optools.FuncOpWrapper;
-import hat.optools.InvokeOpWrapper;
+import hat.optools.OpTk;
 import hat.optools.StructuralOpWrapper;
 import hat.util.StreamCounter;
 import jdk.incubator.code.dialect.java.ClassType;
@@ -155,7 +155,7 @@ public abstract class C99HATKernelBuilder<T extends C99HATKernelBuilder<T>> exte
 
     @Override
     public T type(HATCodeBuilderContext buildContext, JavaType javaType) {
-        if (InvokeOpWrapper.isAssignable(buildContext.lookup,javaType, MappableIface.class) && javaType instanceof ClassType classType){
+        if (OpTk.isAssignable(buildContext.lookup,javaType, MappableIface.class) && javaType instanceof ClassType classType){
               //  .isIfaceUsingLookup(buildContext.lookup,javaType) && javaType instanceof ClassType classType) {
             globalPtrPrefix().space();
             String name = classType.toClassName();
@@ -186,7 +186,7 @@ public abstract class C99HATKernelBuilder<T extends C99HATKernelBuilder<T>> exte
         HATCodeBuilderContext buildContext = new HATCodeBuilderContext(kernelReachableResolvedMethodCall.funcOpWrapper().lookup,kernelReachableResolvedMethodCall.funcOpWrapper());
         buildContext.scope(buildContext.funcOpWrapper, () -> {
             nl();
-            functionDeclaration(buildContext,buildContext.funcOpWrapper.getReturnType(), buildContext.funcOpWrapper.functionName());
+            functionDeclaration(buildContext,(JavaType) buildContext.funcOpWrapper.op.body().yieldType(), buildContext.funcOpWrapper.functionName());
 
             var list = buildContext.funcOpWrapper.paramTable.list();
             parenNlIndented(_ ->
@@ -194,7 +194,7 @@ public abstract class C99HATKernelBuilder<T extends C99HATKernelBuilder<T>> exte
             );
 
             braceNlIndented(_ -> {
-                StreamCounter.of(buildContext.funcOpWrapper.wrappedRootOpStream(), (c, root) ->
+                StreamCounter.of(OpTk.wrappedRootOpStream(buildContext.lookup,buildContext.funcOpWrapper.op), (c, root) ->
                         nlIf(c.isNotFirst()).recurse(buildContext, root).semicolonIf(!(root instanceof StructuralOpWrapper<?>))
                 );
             });
@@ -207,7 +207,7 @@ public abstract class C99HATKernelBuilder<T extends C99HATKernelBuilder<T>> exte
         HATCodeBuilderContext buildContext = new HATCodeBuilderContext(funcOpWrapper.lookup,funcOpWrapper);
         buildContext.scope(buildContext.funcOpWrapper, () -> {
             nl();
-            functionDeclaration(buildContext,buildContext.funcOpWrapper.getReturnType(), buildContext.funcOpWrapper.functionName());
+            functionDeclaration(buildContext,(JavaType) buildContext.funcOpWrapper.op.body().yieldType(), buildContext.funcOpWrapper.functionName());
 
             var list = buildContext.funcOpWrapper.paramTable.list();
             parenNlIndented(_ ->
@@ -215,7 +215,7 @@ public abstract class C99HATKernelBuilder<T extends C99HATKernelBuilder<T>> exte
             );
 
             braceNlIndented(_ -> {
-                StreamCounter.of(buildContext.funcOpWrapper.wrappedRootOpStream(), (c, root) ->
+                StreamCounter.of(OpTk.wrappedRootOpStream(buildContext.lookup,buildContext.funcOpWrapper.op), (c, root) ->
                         nlIf(c.isNotFirst()).recurse(buildContext, root).semicolonIf(!(root instanceof StructuralOpWrapper<?>))
                 );
             });
@@ -224,12 +224,9 @@ public abstract class C99HATKernelBuilder<T extends C99HATKernelBuilder<T>> exte
     }
 
     public T kernelEntrypoint(KernelEntrypoint kernelEntrypoint,Object... args) {
-
         nl();
         HATCodeBuilderContext buildContext = new HATCodeBuilderContext(kernelEntrypoint.funcOpWrapper().lookup,kernelEntrypoint.funcOpWrapper());
-        //  System.out.print(kernelReachableResolvedMethodCall.funcOpWrapper().toText());
         buildContext.scope(buildContext.funcOpWrapper, () -> {
-
             kernelDeclaration(buildContext.funcOpWrapper.functionName());
             // We skip the first arg which was KernelContext.
             var list = buildContext.funcOpWrapper.paramTable.list();
@@ -250,7 +247,7 @@ public abstract class C99HATKernelBuilder<T extends C99HATKernelBuilder<T>> exte
 
             braceNlIndented(_ -> {
                 scope();
-                StreamCounter.of(buildContext.funcOpWrapper.wrappedRootOpStream(), (c, root) ->
+                StreamCounter.of(OpTk.wrappedRootOpStream(buildContext.lookup,buildContext.funcOpWrapper.op), (c, root) ->
                         nlIf(c.isNotFirst()).recurse(buildContext, root).semicolonIf(!(root instanceof StructuralOpWrapper<?>))
                 );
             });
