+++ hat/core/src/main/java/hat/optools/BinaryOpWrapper.java
@@ -33,13 +33,5 @@ public abstract class BinaryOpWrapper<T extends Op> extends OpWrapper<T> {
         super(lookup, op);
     }
 
-    public static Op lhsAsOp(Op op) {
-        return ((Op.Result)op.operands().getFirst()).op();
-    }
-
-    public static Op rhsAsOp(Op op) {
-        return  ((Op.Result)op.operands().get(1)).op();
-    }
-
 
 }
