+++ hat/core/src/main/java/hat/optools/FieldLoadOpWrapper.java
@@ -24,27 +24,12 @@
  */
 package hat.optools;
 
-import jdk.incubator.code.dialect.java.ClassType;
 import jdk.incubator.code.dialect.java.JavaOp;
 
 import java.lang.invoke.MethodHandles;
-import java.lang.reflect.Field;
 
 public class FieldLoadOpWrapper extends FieldAccessOpWrapper<JavaOp.FieldAccessOp.FieldLoadOp> implements LoadOpWrapper {
     FieldLoadOpWrapper( MethodHandles.Lookup lookup,JavaOp.FieldAccessOp.FieldLoadOp op) {
         super(lookup,op);
-    }
-    public static Object getStaticFinalPrimitiveValue(MethodHandles.Lookup lookup,JavaOp.FieldAccessOp.FieldLoadOp fieldLoadOp) {
-        if (fieldType(fieldLoadOp) instanceof ClassType classType) {
-            Class<?> clazz = (Class<?>) classTypeToType(lookup,classType);
-            try {
-                Field field = clazz.getField(fieldName(fieldLoadOp));
-                field.setAccessible(true);
-                return field.get(null);
-            } catch (NoSuchFieldException | IllegalAccessException e) {
-                throw new RuntimeException(e);
-            }
-        }
-        throw new RuntimeException("Could not find field value" + fieldLoadOp);
     }
 }
