+++ hat/core/src/main/java/hat/optools/ForOpWrapper.java
@@ -24,37 +24,13 @@
  */
 package hat.optools;
 
-import jdk.incubator.code.dialect.core.CoreOp;
 import jdk.incubator.code.dialect.java.JavaOp;
 
 import java.lang.invoke.MethodHandles;
-import java.util.ArrayList;
 import java.util.stream.Stream;
 
 public class ForOpWrapper extends LoopOpWrapper<JavaOp.ForOp> {
     ForOpWrapper( MethodHandles.Lookup lookup,JavaOp.ForOp op) {
         super(lookup,op);
-    }
-
-    public Stream<OpWrapper<?>> initWrappedYieldOpStream() {
-       return  op.init().entryBlock().ops().stream().filter(o->o instanceof CoreOp.YieldOp).map(o->OpWrapper.wrap(lookup,o) );
-    }
-
-    @Override
-    public Stream<OpWrapper<?>> conditionWrappedYieldOpStream() {
-        return  op.cond().entryBlock().ops().stream().filter(o->o instanceof CoreOp.YieldOp).map(o->OpWrapper.wrap(lookup,o) );
-    }
-
-    public Stream<OpWrapper<?>> mutateRootWrappedOpStream() {
-          return RootSet.rootsWithoutVarFuncDeclarationsOrYields(lookup,op.bodies().get(2).entryBlock());
-    }
-
-    @Override
-    public Stream<OpWrapper<?>> loopWrappedRootOpStream() {
-        var list = new ArrayList<>(RootSet.rootsWithoutVarFuncDeclarationsOrYields(lookup,op.bodies().get(3).entryBlock()).toList());
-        if (list.getLast() instanceof JavaContinueOpWrapper) {
-            list.removeLast();
-        }
-        return list.stream();
     }
 }
