+++ hat/core/src/main/java/hat/optools/IfOpWrapper.java
@@ -33,7 +33,4 @@ public class IfOpWrapper extends StructuralOpWrapper<JavaOp.IfOp> {
         super(lookup,op);
     }
 
-    public static boolean hasElse(JavaOp.IfOp op,  int idx) {
-        return op.bodies().size()>idx && op.bodies().get(idx).entryBlock().ops().size() > 1;
-    }
 }
