+++ hat/core/src/main/java/hat/optools/ModuleOpWrapper.java
@@ -25,89 +25,13 @@
 package hat.optools;
 
 import java.lang.invoke.MethodHandles;
-import java.lang.reflect.Method;
 
-import hat.callgraph.CallGraph;
-import jdk.incubator.code.Op;
 import jdk.incubator.code.dialect.core.CoreOp;
-import jdk.incubator.code.dialect.java.JavaOp;
-import jdk.incubator.code.dialect.java.MethodRef;
 
 import java.util.*;
 
 public class ModuleOpWrapper extends OpWrapper<CoreOp.ModuleOp> {
     public ModuleOpWrapper(MethodHandles.Lookup lookup, CoreOp.ModuleOp op) {
         super(lookup,op);
-    }
-
-    public SequencedMap<String, CoreOp.FuncOp> functionTable() {
-        return op.functionTable();
-    }
-
-    public static CoreOp.ModuleOp createTransitiveInvokeModule(MethodHandles.Lookup l,
-                                                        FuncOpWrapper entry, CallGraph<?> callGraph) {
-        LinkedHashSet<MethodRef> funcsVisited = new LinkedHashSet<>();
-        List<CoreOp.FuncOp> funcs = new ArrayList<>();
-        record RefAndFunc(MethodRef r, FuncOpWrapper f) {}
-
-        Deque<RefAndFunc> work = new ArrayDeque<>();
-        entry.op.traverse(null, (map, op) -> {
-            if (op instanceof JavaOp.InvokeOp invokeOp) {
-                var invokeOpWrapper = (InvokeOpWrapper) OpWrapper.wrap(entry.lookup,invokeOp);
-                MethodRef methodRef = invokeOpWrapper.methodRef();
-                Method method = null;
-                Class<?> javaRefTypeClass = invokeOpWrapper.javaRefClass().orElseThrow();
-                try {
-                    method = methodRef.resolveToMethod(l, invokeOpWrapper.op.invokeKind());
-                } catch (ReflectiveOperationException _) {
-                    throw new IllegalStateException("Could not resolve invokeWrapper to method");
-                }
-                Optional<CoreOp.FuncOp> f = Op.ofMethod(method);
-                if (f.isPresent() && !callGraph.filterCalls(f.get(), invokeOpWrapper, method, methodRef, javaRefTypeClass)) {
-                    work.push(new RefAndFunc(methodRef, new FuncOpWrapper(l, f.get())));
-                }
-              //  consumer.accept(wrap(lookup,invokeOp));
-            }
-            return map;
-        });
-
-        while (!work.isEmpty()) {
-            RefAndFunc rf = work.pop();
-            if (!funcsVisited.add(rf.r)) {
-                continue;
-            }
-
-            CoreOp.FuncOp tf = rf.f.op.transform(rf.r.name(), (blockBuilder, op) -> {
-                if (op instanceof JavaOp.InvokeOp iop) {
-                    InvokeOpWrapper iopWrapper = OpWrapper.wrap(entry.lookup, iop);
-                    MethodRef methodRef = iopWrapper.methodRef();
-                    Method invokeOpCalledMethod = null;
-                    try {
-                        invokeOpCalledMethod = methodRef.resolveToMethod(l, iop.invokeKind());
-                    } catch (ReflectiveOperationException _) {
-                        throw new IllegalStateException("Could not resolve invokeWrapper to method");
-                    }
-                    if (invokeOpCalledMethod instanceof Method m) {
-                        Optional<CoreOp.FuncOp> f = Op.ofMethod(m);
-                        if (f.isPresent()) {
-                            RefAndFunc call = new RefAndFunc(methodRef, new FuncOpWrapper(l, f.get()));
-                            work.push(call);
-
-                            Op.Result result = blockBuilder.op(CoreOp.funcCall(
-                                    call.r.name(),
-                                    call.f.op.invokableType(),
-                                    blockBuilder.context().getValues(iop.operands())));
-                            blockBuilder.context().mapValue(op.result(), result);
-                            return blockBuilder;
-                        }
-                    }
-                }
-                blockBuilder.op(op);
-                return blockBuilder;
-            });
-            funcs.addFirst(tf);
-        }
-
-        return CoreOp.module(funcs);
     }
 }
