+++ hat/core/src/main/java/hat/optools/VarOpWrapper.java
@@ -25,21 +25,12 @@
 package hat.optools;
 
 import jdk.incubator.code.dialect.core.CoreOp;
-import jdk.incubator.code.dialect.java.JavaType;
 
 import java.lang.invoke.MethodHandles;
 
 public abstract class VarOpWrapper extends OpWrapper<CoreOp.VarOp> {
     public VarOpWrapper(MethodHandles.Lookup lookup,CoreOp.VarOp op) {
         super(lookup, op);
-    }
-
-    public static JavaType javaType(CoreOp.VarOp op) {
-        return (JavaType) op.varValueType();
-    }
-
-    public static String varName(CoreOp.VarOp op) {
-        return op.varName();
     }
 
 }
