+++ hat/tools/src/main/java/hat/tools/text/JavaHATCodeBuilder.java
@@ -26,18 +26,16 @@ package hat.tools.text;
 
 import hat.codebuilders.HATCodeBuilderContext;
 import hat.codebuilders.HATCodeBuilderWithContext;
-import hat.optools.FieldAccessOpWrapper;
 import hat.optools.FieldLoadOpWrapper;
 import hat.optools.FuncOpWrapper;
 import hat.optools.InvokeOpWrapper;
+import hat.optools.OpTk;
 import hat.optools.OpWrapper;
 import hat.optools.StructuralOpWrapper;
 import jdk.incubator.code.Op;
-import jdk.incubator.code.dialect.core.CoreOp;
 import jdk.incubator.code.dialect.java.JavaType;
 import jdk.incubator.code.dialect.java.PrimitiveType;
 
-import java.lang.invoke.MethodHandle;
 import java.lang.invoke.MethodHandles;
 
 public  class JavaHATCodeBuilder<T extends JavaHATCodeBuilder<T>> extends HATCodeBuilderWithContext<T> {
@@ -53,10 +51,10 @@ public  class JavaHATCodeBuilder<T extends JavaHATCodeBuilder<T>> extends HATCod
 
     @Override
     public T fieldLoad(HATCodeBuilderContext buildContext, FieldLoadOpWrapper fieldLoadOpWrapper) {
-        if (FieldAccessOpWrapper.isKernelContextAccess(fieldLoadOpWrapper.op)) {
-            identifier("kc").dot().identifier(FieldAccessOpWrapper.fieldName(fieldLoadOpWrapper.op));
+        if (OpTk.isKernelContextAccess(fieldLoadOpWrapper.op)) {
+            identifier("kc").dot().identifier(OpTk.fieldName(fieldLoadOpWrapper.op));
         } else if (fieldLoadOpWrapper.op.operands().isEmpty() && fieldLoadOpWrapper.op.result().type() instanceof PrimitiveType) { // only primitve fields
-            var value = FieldLoadOpWrapper.getStaticFinalPrimitiveValue(buildContext.lookup,fieldLoadOpWrapper.op);
+            var value = OpTk.getStaticFinalPrimitiveValue(buildContext.lookup,fieldLoadOpWrapper.op);
             literal(value.toString());
         } else {
             throw new IllegalStateException("An instance field? I guess - we dont get those in HAT " + fieldLoadOpWrapper.op);
@@ -84,10 +82,11 @@ public  class JavaHATCodeBuilder<T extends JavaHATCodeBuilder<T>> extends HATCod
                 commaSeparated(funcOpWrapper.paramTable.list(), (info) -> type(buildContext,(JavaType) info.parameter.type()).space().varName(info.varOp))
         );
         braceNlIndented(_ ->
-                funcOpWrapper.wrappedRootOpStream()
+                OpTk.wrappedRootOpStream(buildContext.lookup,funcOpWrapper.op)
                         .forEach(root ->
-                        recurse(buildContext, root).semicolonIf(!(root instanceof StructuralOpWrapper<?>)).nl()
-                ));
+                                recurse(buildContext, root).semicolonIf(!(root instanceof StructuralOpWrapper<?>)).nl()
+                        )
+        );
         return self();
     }
